Created ClockTesting, which is used for controlled scheduling and execution of tasks during testing

+++ src/async/TimerAsyncScheduler.h
@@ -11,12 +11,12 @@ namespace async {
 class TimerAsyncScheduler {
  public:
   // Schedule the callback to run once in the background after `wait` amount of time.
-  virtual void schedule_once(std::function<void(void)> callback, unsigned wait) = 0;
+  virtual void schedule_once(std::function<void(void)> callback, long wait) = 0;
 
   // Schedul the callback to run repeatedly with period `period`. Note that
   // the first time the callback is run after `wait` amount of time; it doesn't
   // execute immediately.
-  virtual void schedule_repeated(std::function<void(void)> callback, unsigned period) = 0;
+  virtual void schedule_repeated(std::function<void(void)> callback, long period) = 0;
 };
 
 } // namespace async
